Extract `PDFPageViewBase` class out of `PDFPageView`
This base class contains the generic logic for:
- Creating a canvas and showing when appropriate
- Rendering in the canvas
- Keeping track of the rendering state
